+++ java/utils.scala
@@ -352,11 +352,12 @@ object Utils {
   def addSequentialIndex(
     sc: SparkContext,
     df: DataFrame,
+    from: Int,
     id: String) : DataFrame = {
       val sqlContext = new org.apache.spark.sql.SQLContext(sc)
       sqlContext.createDataFrame(
         df.rdd.zipWithIndex.map {
-          case (row: Row, i: Long) => Row.fromSeq(row.toSeq :+ i.toDouble)
+          case (row: Row, i: Long) => Row.fromSeq(row.toSeq :+ (i.toDouble + from.toDouble))
         },
       df.schema.add(id, "double")
       )